4주차 실습 과제 (미완성)
