+++ server/routes/userRoutes.js
@@ -1,0 +1,1 @@
+// 사용자 관련 라우팅
